pyproject.toml
[project]
name = "cmsi638-stocktracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.2.0",
    "google-adk>=1.28.1",
    "google-genai>=1.70.0",
    "massive>=2.4.0",
    "mcp>=1.27.0",
    "polygon>=1.2.8",
    "polygon-api-client>=1.16.3",
    "python-dotenv>=1.2.2",
]
